1.0.0
massive rewrite to the whole architecture. Most of the old commands are
not compatible anymore, but the infrastructure is guaranteed to be 247%
more object oriented! Also, I am trying to learn about the new language
feature of ES2017, so rewriting the bot for like the 3 time seems to be
the best way :D

The main change here is to write MeganeBot as an API that encapsulate the
discord.client, instead of just interfacing with it. This allows for
better modulatity in command definition, and allows for cleaner code by
abstrating a lot of the features into their own objects.

There are actual classes for commands, messages, modules and a dispatcher
to connect them all.

There are still many things to do to get this infastructure to work the
way I want, mainly to take away a lot of ambious processing from the util
classe, and create type-safe arguements for commands, and create better
regex parsing for command strings instead of substring parsing.

Lastly, I would like to fully disclose that I am basing this design upon
the Commando command framework for discord.js, since I wanted to separate
the argument parsing and command management away from the commands
themselves.